requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
peewee==3.17.7
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.1.3
